changes added

+++ controllers/doctor.controller.js
@@ -31,7 +31,7 @@ module.exports.doctorLogin = async (req) => {
         // console.log("password",password, req.doctor_password );
         if (password == req.doctor_password) {
             if (moment().isBefore(data[0].subscription_end_date)) {
-                return ({ status: true, data: data });
+                return ({ status: true, data: data, msg: 'Login Successful !!!' });
             } else {
                 return ({ status: false, msg: 'Your Subscription Has Expired' });
             }
@@ -101,16 +101,12 @@ exports.deleteDoctor = async (req, res) => {
     }
 }
 
-// async function saveFile(){
-//     file.filepath = path.join(__dirname, '../public/uploads/images/') + fileName;
-// }
-
 module.exports.addDoctor = async (req, res) => {
     let bodyObj = {}
-    var form = await new formidable.IncomingForm();
+    var form = new formidable.IncomingForm();
     form.parse(req);
 
-    form.on('fileBegin', async function (name, file) {
+    form.on('fileBegin', function (name, file) {
         if (file.originalFilename != '' && file.originalFilename != undefined && file.originalFilename != null) {
             if (file.mimetype && file.mimetype.includes('image')) {
                 let ext = file.originalFilename.split('.')[1];
@@ -123,46 +119,43 @@ module.exports.addDoctor = async (req, res) => {
 
     form.on('field', function (name, value) {
         bodyObj[name] = value;
-        console.log("bodyObj", name, value);
     });
-
     form.on('error', (err) => {
-        return ({ status: false, err: err })
+        res.send({ status: false, err: err })
     });
-
 
     return new Promise(function (resolve, reject) {
         form.on('end', async function () {
-            console.log("i am heree");
             let validation = await validateAddRequest(bodyObj);
             console.log("validation", validation);
             if (!validation.status) {
-                
+                if (!res.headersSent) {
                     if (bodyObj.doctor_logo) await deleteFile(bodyObj.doctor_logo);
 
                     if (bodyObj.doctor_sign) await deleteFile(bodyObj.doctor_sign);
 
-                    resolve(validation)
-                
+
+                    res.send(validation)
+                }
             } else {
-                
+                if (!res.headersSent) {
 
                     bodyObj.doctor_password = await encrypt_decrypt.encrypt(bodyObj.doctor_password)
                     bodyObj.doctor_mobile = parseInt(bodyObj.doctor_mobile)
                     let subscriptionObject = await processSubscription(bodyObj.subscription)
                     let query = `INSERT INTO mst_doctors
-            (doctor_name, doctor_username, doctor_email, doctor_mobile, doctor_password, doctor_degree, specialisation, doctor_logo, doctor_sign, subscription_type,
-            subscription_start_date, subscription_end_date, doctor_address, registration_number, consultation_charge, treatment1_charge, treatment2_charge, treatment3_charge, isActive) 
-            VALUES 
-            (?,?,?,?,?,?,?,?,?,?,
-             ?,?,?,?,?,?,?,?,?)`;//19
+                                (doctor_name, doctor_username, doctor_address, doctor_email, doctor_mobile, doctor_password, doctor_degree, specialisation, doctor_logo, doctor_sign, 
+                                subscription_type, subscription_start_date, subscription_end_date, doctor_address, registration_number, consultation_charge, treatment1_charge, treatment2_charge, treatment3_charge, isActive) 
+                                VALUES 
+                                (?,?,?,?,?,?,?,?,?,?,
+                                 ?,?,?,?,?,?,?,?,?,?)`;//20
 
                     let values = [
-                        bodyObj.doctor_name, bodyObj.doctor_username, bodyObj.doctor_email, bodyObj.doctor_mobile, bodyObj.doctor_password, bodyObj.doctor_degree, bodyObj.specialisation, bodyObj.doctor_logo, bodyObj.doctor_sign, subscriptionObject.subscription_type,
+                        bodyObj.doctor_name, bodyObj.doctor_username, bodyObj.doctor_address, bodyObj.doctor_email, bodyObj.doctor_mobile, bodyObj.doctor_password, bodyObj.doctor_degree, bodyObj.specialisation, bodyObj.doctor_logo, bodyObj.doctor_sign, subscriptionObject.subscription_type,
                         subscriptionObject.subscription_start_date, subscriptionObject.subscription_end_date, bodyObj.doctor_address, bodyObj.registration_number, parseInt(bodyObj.consultation_charge), parseInt(bodyObj.treatment1_charge), parseInt(bodyObj.treatment2_charge), parseInt(bodyObj.treatment3_charge), 1
-                    ]//19
+                    ]//20
+
                     let result = await db.executevaluesquery(query, values)
-                    console.log("add doctor res", result);
                     if (result.insertId) {
                         resolve({ status: true, msg: "Data inserted successfully" })
                     } else {
@@ -177,7 +170,7 @@ module.exports.addDoctor = async (req, res) => {
 
                 }
 
-            
+            }
         })
     })
 
@@ -187,66 +180,6 @@ module.exports.addDoctor = async (req, res) => {
 
 }
 
-
-// async function validateAddRequest(req) {
-//     if (!req.doctor_username) {
-//         return { status: false, msg: 'Please Enter Username' };
-//     }
-
-//     if (req.doctor_username && await checkUsername(req.doctor_username)) {
-//         return { status: false, msg: 'Username Already Exists' };
-//     }
-
-//     if (!req.doctor_password) {
-//         return { status: false, msg: 'Please Enter Password' };
-//     }
-
-//     if (!req.doctor_email) {
-//         return { status: false, msg: 'Please Enter Email' };
-//     }
-
-//     if (req.doctor_email && req.doctor_email.trim().match(/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/)) {
-//         return { status: false, msg: 'Please Enter Valid Email' };
-//     }
-
-//     if (!req.doctor_name) {
-//         return { status: false, msg: 'Please Enter Name' };
-//     }
-
-//     if (!req.doctor_mobile) {
-//         return { status: false, msg: 'Please Enter Mobile Number' };
-//     }
-
-//     if (!req.subscription) {
-//         return { status: false, msg: 'Please Select a Subscription Plan' };
-//     }
-
-//     if (!req.doctor_logo) {
-//         return { status: false, msg: 'Please Select a Logo' };
-//     }
-
-//     if (!req.doctor_sign) {
-//         return { status: false, msg: 'Please Select a Sign' };
-//     }
-
-//     if (!req.consultation_charge) {
-//         return { status: false, msg: 'Please Enter Consultation Charges' };
-//     }
-
-//     if (!req.treatment1_charge) {
-//         return { status: false, msg: 'Please Enter Treatment Type 1 Charges' };
-//     }
-
-//     if (!req.treatment2_charge) {
-//         return { status: false, msg: 'Please Enter Treatment Type 2 Charges' };
-//     }
-
-//     if (!req.treatment3_charge) {
-//         return { status: false, msg: 'Please Enter Treatment Type 3 Charges' };
-//     }
-
-//     return { status: true };
-// }
 
 async function processSubscription(subscription) {
     let arr = subscription.split("-")
